fix: my favorite page
