--- frontend/src/app/(account-management)/my-profile/lib/interface/profile.interface.ts
@@ -1,10 +1,0 @@
-import { ChangeEvent } from "react";
-
-export interface ProfileInfoProps {
-    form: {
-        name: string;
-        email: string;
-        dob: string;
-    };
-    handleChange: (e: ChangeEvent<HTMLInputElement>) => void;
-}
